Connections and organization
Connected the contact, produce, and login; signup, admin login pages to the home pages and routed them properly.

Organized the html files received from Vanessa and Derrick and organized them in the views folder. All the images were put in the public > images folder so that they can be accessed and rerouted them in the <img> tags on those pages